Updated to allow tracking while not projecting as well as more refactor
Made it so that instead of stopping tracking when turn off projection it just changes texture to a pure black texture. The slowest part of the kinect is picking up people so this allows the kinect to always be tracking and be able to turn on in an instant. 

Also started renaming variables and cleaning up some of the unneeded commented code.